Refactor Docker setup and composer installation
This commit consists of significant Docker setup refactoring and the addition of composer installation into Dockerfile. The php-fpm Dockerfile was deleted, and its setup was moved into the main Dockerfile with the inclusion of composer installation. Docker-compose and .gitignore files were updated respectively, along with the renaming of 'nginx/nginx.conf' and 'php-fpm/php-ini-overrides.ini' for simplified addressing.